requirements.txt
av==12.1.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
ctranslate2==4.3.1
faster-whisper==1.0.2
ffmpeg-python==0.2.0
filelock==3.15.1
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2024.6.0
future==1.0.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
numpy==2.0.0
onnxruntime==1.18.0
packaging==24.1
protobuf==5.27.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.0.0
sympy==1.12.1
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
